Isolated SMCRA tests
